Resolved some coding standard issue in the modules and themes file
