+++ web/modules/custom/profile_block/profile_block.module
@@ -1,34 +1,42 @@
 <?php
- 
+
+/**
+ * @file
+ * To store the variables that are been displayed on the profile page.
+ */
+
+/**
+ * To store the variable in the user profile block.
+ */
 function profile_block_theme($existing, $type, $theme, $path) {
   return [
 
     'user-profile-block' => [
-      'variables' => [ 
-          'uname' => NULL,  
-          'post_count' => NULL,
-          'followers_count' => NULL,
-          'following_count' => NULL,  
-          'name' => NULL, 
-          'bio' => NULL, 
-          'markup1' => NULL,
-          'markup2' => NULL,
-          'attached' => NULL,
+      'variables' => [
+        'uname' => NULL,
+        'post_count' => NULL,
+        'followers_count' => NULL,
+        'following_count' => NULL,
+        'name' => NULL,
+        'bio' => NULL,
+        'markup1' => NULL,
+        'markup2' => NULL,
+        'attached' => NULL,
       ],
     ],
 
     'flag-modal' => [
-      'variables' => [   
-        'uri' => NULL,  
-        'uname' => NULL, 
-        'name' => NULL,  
-        'post_count' => NULL, 
-        'followers_count' => NULL, 
-        'following_count' => NULL, 
-        'post_uri_all' => NULL,  
-        'link' => NULL, 
-        'flag_txt' => NULL,  
-        'flag_link' => NULL,   
+      'variables' => [
+        'uri' => NULL,
+        'uname' => NULL,
+        'name' => NULL,
+        'post_count' => NULL,
+        'followers_count' => NULL,
+        'following_count' => NULL,
+        'post_uri_all' => NULL,
+        'link' => NULL,
+        'flag_txt' => NULL,
+        'flag_link' => NULL,
       ],
     ],
   ];
