+++ web/modules/custom/profile_block/profile_block.services.yml
@@ -1,5 +1,4 @@
 services:
       profile_block.CustomModalController:
         class: Drupal\profile_block\Controller\CustomModalController
-        arguments: ['@database'] 
-        
+        arguments: ['@database']
